[usbdev_stream_test] Refactoring and renaming
Migrate the stream handling code into lib/testing to form
part of the usb_testutils code. This will allow inclusion
in additional top-level usbdev tests. Rename subsequently.
